requirements.txt
pystray>=0.19.0
pillow>=10.0.0
pyinstaller>=6.0.0
requests>=2.31.0